ct hackathon
KrishiMitra is an AI-enabled agricultural platform that provides farmers with real-time weather forecasts and farmer market prices, soil analysis and recommendations for crops, advice on planting and care, as well as a multi-modal chatbot for 24/7 agricultural support, in up to six languages! Farmers have access to their soil and weather analyses, potential prices and pest pressure, and links directly to farmers' markets and suppliers. The AI and machine learning platform uses integration of external APIs, and is largely focused on productivity, building stability of income, and ensuring sustainable, healthy farming practices!